Legal Ease Final/legalease_v2/legalease/legalease-python-backend/requirements.txt
# ---------------------------------------------------------------------------
# LegalEase AI — Python backend dependencies
# ---------------------------------------------------------------------------

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart>=0.0.18       # multipart/form-data (file uploads)

# Config
python-dotenv==1.0.1

# Supabase (DB + Storage)
supabase==2.7.4

# Google Gemini (chat, embeddings, TTS, STT)
google-generativeai==0.8.3

# PDF text extraction
pypdf==4.3.1

# HTTP client (used for Gemini TTS REST call)
httpx==0.27.2

# Text-to-speech (Indian language support via Google Translate TTS)
gTTS==2.5.3
